Update fluctuations on primary ionization and attenuation
This commit doesn't significantly change the computation time or the energy resolution.

We are introducing Poisson fluctuations in the number of primary electrons produced during ionization. Note: for very low-energy tracks, we should find a way to constrain the total number of primary electrons produced. In fact, with Poisson fluctuations, there's a possibility of getting no electrons at all.

Additionally, we are replacing the Poisson fluctuation in attenuation with a binomial distribution.